change viewport type

+++ Examples/RetroExample/retro_example_app.cpp
@@ -962,8 +962,8 @@ void RetroExampleApp::render()
         .left = 0,
         .top = 0,
         .front = 0,
-        .right = static_cast<long>(dim.x),
-        .bottom = static_cast<long>(dim.y),
+        .right = dim.x,
+        .bottom = dim.y,
         .back = 0,
     };
     m_context->set_viewports(&cmd_list, 1, &viewport);
@@ -1137,8 +1137,8 @@ void RetroExampleApp::render()
         .left = 0,
         .top = 0,
         .front = 0,
-        .right = static_cast<long>(bloom_w),
-        .bottom = static_cast<long>(bloom_h),
+        .right = bloom_w,
+        .bottom = bloom_h,
         .back = 0,
     };
     m_context->set_viewports(&cmd_list, 1, &bloom_viewport);
